data source logic refactor
