--- client_test.py
@@ -1,3 +1,0 @@
-from client_data.client import *
-
-add_transactions('tfsa')
