--- data_sources/data_source_manager.py
@@ -1,53 +1,0 @@
-from data_sources.alphavantage_connection import AlphaVantageConnection
-from data_sources.simfin_connection import SimFinConnection
-from config.config_utils import fetch_data_sources
-
-
-class DataSourceManager(object):
-    _prices_data_source = AlphaVantageConnection()
-    _statements_data_source = AlphaVantageConnection()
-    NAME = 'Data Source Manager'
-
-    def __init__(self):
-        self.set_sources()
-
-    def __repr__(self):
-        return self.NAME
-
-    def set_sources(self):
-        prices_data_source = fetch_data_sources('prices')
-        statements_data_source = fetch_data_sources('statements')
-        if prices_data_source == 'AlphaVantage':
-            prices = AlphaVantageConnection()
-        elif prices_data_source == 'SimFin':
-            prices = SimFinConnection()
-        else:
-            raise ValueError(f'prices datasource: {prices_data_source} not valid')
-
-        if statements_data_source == 'AlphaVantage':
-            statements = AlphaVantageConnection()
-        elif statements_data_source == 'SimFin':
-            statements = SimFinConnection()
-        else:
-            raise ValueError(f'statements datasource: {statements_data_source} not valid')
-
-        self._prices_data_source = prices
-        self._statements_data_source = statements
-        return prices, statements
-
-    def prices_local(self,
-                     ticker: str,
-                     read: bool = True):
-        return self._prices_data_source.get_prices(ticker=ticker, read=read)
-
-    def fx(self,
-           currency: str,
-           read: bool = True):
-        return self._prices_data_source.get_fx(currency=currency, read=read)
-
-    def refresh_fx(self,
-                   currency: str) -> None:
-        self._prices_data_source.get_fx(currency=currency, read=False)
-
-    def statements(self):
-        raise NotImplemented
